Rename user_counter folder to userCounter
