+++ app/packs/src/apps/userCounter/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import UserCounter from 'src/apps/user_counter/UserCounter';
+import UserCounter from 'src/apps/userCounter/UserCounter';
 
 document.addEventListener('DOMContentLoaded', () => {
   const domElement = document.getElementById('UserCounter');
